# minor pkg/ip fixes