A few small drive-by improvements to `pkg/ip`.


```
ip: remove unneeded IPTestSuite

Likely left over from a conversion from checkmate to go testing.
```
```
ip: benchmark SortAddrList

To create a baseline for the next commit.
```
```
ip: use stdlib slices package

This gives us a speedup of 30% in a microbenchmark for sorting IPs and
moves us closer to stdlib.

pkg: github.com/cilium/cilium/pkg/ip
                │     pre      │                post                 │
                │    sec/op    │   sec/op     vs base                │
SortAddrList-10   105.06µ ± 4%   72.75µ ± 1%  -30.75% (p=0.000 n=10)
```
```
slices: drop SortedUniqueFunc

The type constraints of slices.Sort and slices.Compact are different -
the former requires cmp.Ordered whereas the latter only needs
comparable. The only caller for SortedUniqueFunc was pkg/ip, for
[]netip.Addr. The usage is inefficient, though, since netip.Addr is
comparable, just not cmp.Ordered - the use of slices.CompactFunc is thus
unnecessary.

Since there's no consumers, and use can be needlessly inefficient, let's
just drop the function, even though it breaks API symmetry.
```